quitdwm: Only kill one dwm per call
Perviously, all instances of dwm would be killed when quitdwm was
called, even those on other ttys. Now, we only kill the instance of dwm
on the tty most recently logged into by the user.

This is good enough for what I need in the very few cases where I have more
than one instance of dwm running as the same user. If necessary, the particular
dwm to kill can be determined with `ps a | grep dwm` and choosing the
pid listed with the desired tty.