watchyt: No longer limit video dimensions to screen dimensions
There were a few reasons for this change.

For one thing, it turns out that there *is* a noticable difference in
quality between 720p video and 1080p video, even on a 1280x720 screen.

Secondly, on my 2560x1440 desktop monitor, videos were still being
limited to 720p, even when videos had a 1080p format available as well.
Something was clearly wrong.

Thirdly, the main reason I implemented this in the first place was that
I didn't want to waste bandwidth on 4K video when it wouldn't make that
much of a difference. However, while 1080p video does take longer to
download than 720p, it is not nearly as long as it takes to download 4k
video. For this reason, I have chosen to implement a static cap of 1080p
instead of a dynamic cap based on the screen size.