src/watchyt: Use yt-dlp for streaming video
While there are many differences (improvements?) in yt-dlp compared to
youtube-dl, the one that mattered to me in making this decision is that
at the time of writing youtube-dl seems to be quite severely throttled,
to rates around 64kb/s. Even before this, a perennial issue with this
script is that when loading a video, it would be buffered, and the only
solution would be to close mpv and run the script again. Maybe yt-dlp
will address that issue too.