README.md: Use '#' style headings instead of '==='
I prefer this style of heading because of it is much easier to
parse, for example to format colors as in [0]. To be able to parse
'==='-style headings, we would have to store a line in memory, and not
print it until we read the next line to see if it is a heading. Much
more complicated.

[0]: https://git.sr.ht/~smlavine/scripts/tree/master/item/src/lessdown