src/touch-menu: Check for pen input before rotating screen
If the pen or eraser is not found, then the user (probably just) hasn't
touched it to the screen yet in this X session. This check addresses the
annoying situation where I flip the screen, then for the first time in
the session use the touch pen, which is misaligned, making me waste time
rotating the screen over again.