src/weather: Use network UUIDs instead of SSIDs
Using UUIDs is more robust because it avoids the (unlikely) possibility
of a name collision.

This commit also changes how we call nmcli to obtain this information.
We now use the -g option, which is explicitly tailored to use in scripts
and removes the need for a sed call to filter out the header.

The unnecessary cut call (we're already using awk!) was removed.

The use of e.g. was corrected to i.e.

Note that this is a BREAKING change; you will need to either delete your
weather.tsv and let the script find your network connections again, or
manually convert the SSIDs to UUIDs yourself.