src/parsed-xrandr: Print "active" if active
This will be easier to parse, rather than having to check for
"connected" and also the presence of a resolution.

This also fixes a bug where a connected but inactive input would not be
displayed, because the statement that printed info about connected input
was only in the block where a * indicating the active resolution was found.