src/musicsum: Rewrite filter from awk to jq
This fixes a few problems with the previous approach.

First: The tr and cut calls are unnecessary. awk is already capable of
splitting fields on multiple delimiters, and fields 3 and 5 simply could
have been referenced in the original awk script instead of redefining
them as fields 1 and 2.

Using jq fixes this problem by having mediainfo produce its output in a
structured format, so that this trickery wouldn't be necessary even if
awk couldn't handle it automatically.

Second: The "NR % 2 != 1" is not properly documented and was a bit of a
hack. The reason for this is that mediainfo outputs the duration of a
file twice: once when printing the "General" information about the file,
and again when printing the "Audio" information about the file. This
condition ignored one of them, to avoid the sum being double what was
desired.

Using jq fixes this problem by allowing us to clearly and legibly filter
the input -- in this case, by obviously stating that we are only
interested in the "Audio" information about the file.

Third: The code was subtly incorrect, and based on an incomplete
premise. mediainfo outputs the duration of an audio file on a line like
the following:

Duration                                 : 1 min 40 s

However, if a file has a duration of less than a minute, it will instead
output this:

Duration                                 : 31 s 274 ms

Using jq sidesteps this issue because when mediainfo is prompted to
output as JSON, it provides the duration just as seconds, which is the
information that we originally reconstructed after the fact in awk.

One small downside of jq is that the formatting code is slightly more
awkward, in part because unlike awk jq's type system requires us to
explicitly convert between numbers and strings. However, the output
formatting was already slightly awkward because of the inherent
arithmetic. It is worth it to me for the sake of maintainability and
speed to only use one language for both filtering the input and
formatting the output.